move environement selection logic and ui to TestEnvironmentSelection.vue
